+++ test/boolean_validation_test.dart
@@ -1,2 +1,2 @@
-/// all tests will run automatically from the validation folder
+/// all tests will run automatically from the validations folder
 void main() {}